vmstat: adapted for changes in an i/f, <DISKSTATS> api
This patch is a response to changes to the <diskstats>
interface. And the following represents the summary of
significant unrelated alterations that were also made.

+ corrected that 'milli weighted IO' output, which has
been wrong since that original patch referenced below.
as implemented, it duplicated 'milli spent IO' output.

+ restored original commit intent regarding disks with
a partition switch as represented in references below.

+ moved all item enumerators up near the source start.

+ removed all remaining tabs and inconsistent indents.

+ reformatted the silly style applied to 'xerrx' uses.

Reference(s):
. disallowed 'disks' under 'partition' switch
commit 7df7795b929c2c0911127d9fff64e62fe9e2975f
. original commit with disk/partition rational
commit e445f7e6c5539594600ace7a722dafa3b91a5374

Signed-off-by: Jim Warner <james.warner@comcast.net>